+++ ml_tools/ML_scaler/__init__.py
@@ -2,11 +2,16 @@ from ._ML_scaler import (
     DragonScaler
 )
 
+from ._scaler_handler import (
+    DragonScalerHandler
+)
+
 from .._core import _imprimir_disponibles
 
 
 __all__ = [
-    "DragonScaler"
+    "DragonScaler",
+    "DragonScalerHandler"
 ]