+++ ml_tools/VIF/__init__.py
@@ -1,16 +1,12 @@
-from ._VIF_factor import (
-    compute_vif,
-    drop_vif_based,
-    compute_vif_multi,
+from ._dragon_vif import (
+    DragonVIF
 )
 
 from .._core import _imprimir_disponibles
 
 
 __all__ = [
-    "compute_vif",
-    "drop_vif_based",
-    "compute_vif_multi"
+    "DragonVIF"
 ]